requirements.txt
python-dotenv
requests
openai
apify-client
schedule
tzlocal
pytz
